Architecture module

ControllerFormats.Architecture — Module

Architecture

Module containing data structures to represent controllers.

Architecture module

Neural networks

General interface

Layer operations

Activation functions

Neural networks

An artificial neural network can be used as a controller.

General interface

ControllerFormats.Architecture.AbstractNeuralNetwork — Type

AbstractNeuralNetwork

Abstract type for neural networks.

ControllerFormats.Architecture.dim_in — Method

dim_in(N::AbstractNeuralNetwork)

Return the input dimension of a neural network.

Input

N – neural network

Output

The dimension of the input layer of N.

ControllerFormats.Architecture.dim_out — Method

dim_out(N::AbstractNeuralNetwork)

Return the output dimension of a neural network.

Input

N – neural network

Output

The dimension of the output layer of N.

ControllerFormats.Architecture.dim — Method

dim(N::AbstractNeuralNetwork)

Return the input and output dimension of a neural network.

Input

N – neural network

Output

The pair $(i, o)$ where $i$ is the input dimension and $o$ is the output dimension of N.

Implementation

ControllerFormats.Architecture.FeedforwardNetwork — Type

FeedforwardNetwork{L} <: AbstractNeuralNetwork

Standard implementation of a feedforward neural network which stores the layer operations.

Fields

layers – vector of layer operations (see AbstractLayerOp)

Notes

The field layers contains the layer operations, so the number of layers is length(layers) + 1.

Layer operations

ControllerFormats.Architecture.AbstractLayerOp — Type

AbstractLayerOp

Abstract type for layer operations.

Notes

An AbstractLayerOp represents a layer operation. A classical example is a "dense layer operation" with an affine map followed by an activation function.

ControllerFormats.Architecture.dim_in — Method

dim_in(L::AbstractLayerOp)

Return the input dimension of a layer operation.

Input

L – layer operation

Output

The input dimension of L.

ControllerFormats.Architecture.dim_out — Method

dim_out(L::AbstractLayerOp)

Return the output dimension of a layer operation.

Input

L – layer operation

Output

The output dimension of L.

ControllerFormats.Architecture.dim — Method

dim(L::AbstractLayerOp)

Return the input and output dimension of a layer operation.

Input

N – neural network

Output

The pair $(i, o)$ where $i$ is the input dimension and $o$ is the output dimension of N.

Implementation

ControllerFormats.Architecture.DenseLayerOp — Type

DenseLayerOp{F, M<:AbstractMatrix, B} <: AbstractLayerOp

A dense layer operation is an affine map followed by an activation function.

Fields

weights – weight matrix

bias – bias vector

activation – activation function

Activation functions

ControllerFormats.Architecture.ActivationFunction — Type

ActivationFunction

Abstract type for activation functions.

ControllerFormats.Architecture.Id — Type

Id

Identity activation.

\[ f(x) = x\]

ControllerFormats.Architecture.ReLU — Type

ReLU

Rectified linear unit (ReLU) activation.

\[ f(x) = max(x, 0)\]

ControllerFormats.Architecture.Sigmoid — Type

Sigmoid

Sigmoid activation.

\[ f(x) = \frac{1}{1 + e^{-x}}\]

ControllerFormats.Architecture.Tanh — Type

Tanh

Hyperbolic tangent activation.

\[ f(x) = tanh(x) = \frac{e^x - e^{-x}}{e^x + e^{-x}}\]

The following strings can be parsed as activation functions:

ControllerFormats.FileFormats.available_activations

Dict{String, ActivationFunction} with 11 entries:
  "ReLU"    => ReLU()
  "Affine"  => Id()
  "relu"    => ReLU()
  "Sigmoid" => Sigmoid()
  "Id"      => Id()
  "sigmoid" => Sigmoid()
  "σ"       => Sigmoid()
  "Tanh"    => Tanh()
  "linear"  => Id()
  "tanh"    => Tanh()
  "Linear"  => Id()